add dev parameter

+++ scripts/ValidateCode.sh
@@ -15,23 +15,26 @@ docker-compose exec -T ${CONTAINER_NAME} python -c "import requests; f=open('${C
 [ -d "/mnt/samba/${APP_NAME}" ] || mkdir -p "/mnt/samba/${APP_NAME}"
 docker cp "$(docker-compose ps -q ${CONTAINER_NAME})":"/python/app/${CLEAN_BRANCH_NAME}_openapi.json" "/mnt/samba/${APP_NAME}/${CLEAN_BRANCH_NAME}_openapi.json"
 
-echo "Clean up old reports" 
-rm -f unittesting.xml coverage.xml typing.xml typing-server.xml typing-integrations.xml
+if [[ -n $DEV_ARG ]] 
+then
+    echo "Clean up old reports" 
+    rm -f unittesting.xml coverage.xml typing.xml typing-server.xml typing-integrations.xml
 
-echo "Code tests" 
-## Catch the exit codes so we don't exit the whole script before we are done.
-## Typing, linting, formatting check & unit and integration testing
-if [[ $CONTAINER_NAME != "webapp" ]]; then
-    echo "flake8 tests" #TODO: remove the if else block once the errors on sb-pim are fixed
-    docker-compose exec -T ${CONTAINER_NAME} flake8; STATUS1=$?    # For Sb-pim only
-    
-else
-    docker-compose exec -T ${CONTAINER_NAME} validatecodeonce; STATUS1=$?
-    docker cp "$(docker-compose ps -q ${CONTAINER_NAME})":/python/reports/typing.xml typing.xml
-    docker cp "$(docker-compose ps -q ${CONTAINER_NAME})":/python/reports/unittesting.xml unittesting.xml
-    docker cp "$(docker-compose ps -q ${CONTAINER_NAME})":/python/reports/coverage.xml coverage.xml
-fi
+    echo "Code tests" 
+    ## Catch the exit codes so we don't exit the whole script before we are done.
+    ## Typing, linting, formatting check & unit and integration testing
+    if [[ $CONTAINER_NAME != "webapp" ]]; then
+        echo "flake8 tests" #TODO: remove the if else block once the errors on sb-pim are fixed
+        docker-compose exec -T ${CONTAINER_NAME} flake8; STATUS1=$?    # For Sb-pim only
+        
+    else
+        docker-compose exec -T ${CONTAINER_NAME} validatecodeonce; STATUS1=$?
+        docker cp "$(docker-compose ps -q ${CONTAINER_NAME})":/python/reports/typing.xml typing.xml
+        docker cp "$(docker-compose ps -q ${CONTAINER_NAME})":/python/reports/unittesting.xml unittesting.xml
+        docker cp "$(docker-compose ps -q ${CONTAINER_NAME})":/python/reports/coverage.xml coverage.xml
+    fi
 
-## Return the status code
-TOTAL=$((STATUS1))
-exit $TOTAL
+    ## Return the status code
+    TOTAL=$((STATUS1))
+    exit $TOTAL
+fi
